Move SizedData and AccountData into account_utils
