+++ rust/sealevel/libraries/account-utils/src/lib.rs
@@ -1,3 +1,4 @@
+use borsh::{BorshDeserialize, BorshSerialize};
 use solana_program::{
     account_info::AccountInfo,
     program::{invoke, invoke_signed},
@@ -7,6 +8,129 @@ use solana_program::{
     system_instruction, system_program,
 };
 use spl_type_length_value::discriminator::Discriminator;
+
+/// Data that has a predictable size when serialized.
+pub trait SizedData {
+    /// Returns the size of the data when serialized.
+    fn size(&self) -> usize;
+}
+
+// FIXME should probably define another trait rather than use Default for this as a valid but
+// uninitialized object in rust is a big no no.
+pub trait Data: BorshDeserialize + BorshSerialize + Default {}
+impl<T> Data for T where T: BorshDeserialize + BorshSerialize + Default {}
+
+/// Account data structure wrapper type that handles initialization and (de)serialization.
+///
+/// (De)serialization is done with borsh and the "on-disk" format is as follows:
+/// {
+///     initialized: bool,
+///     data: T,
+/// }
+#[derive(Debug, Default)]
+pub struct AccountData<T> {
+    data: Box<T>,
+}
+
+impl<T> From<T> for AccountData<T> {
+    fn from(data: T) -> Self {
+        Self {
+            data: Box::new(data),
+        }
+    }
+}
+
+impl<T> From<Box<T>> for AccountData<T> {
+    fn from(data: Box<T>) -> Self {
+        Self { data }
+    }
+}
+
+impl<T> SizedData for AccountData<T>
+where
+    T: SizedData,
+{
+    fn size(&self) -> usize {
+        // Add an extra byte for the initialized flag.
+        1 + self.data.size()
+    }
+}
+
+impl<T> AccountData<T>
+where
+    T: Data,
+{
+    pub fn into_inner(self) -> Box<T> {
+        self.data
+    }
+
+    /// Deserializes the account data from the given slice.
+    pub fn fetch_data(buf: &mut &[u8]) -> Result<Option<Box<T>>, ProgramError> {
+        if buf.is_empty() {
+            return Ok(None);
+        }
+        // Account data is zero initialized.
+        let initialized = bool::deserialize(buf)?;
+        let data = if initialized {
+            Some(T::deserialize(buf).map(Box::new)?)
+        } else {
+            None
+        };
+        Ok(data)
+    }
+
+    /// Deserializes the account data from the given slice and wraps it in an `AccountData`.
+    pub fn fetch(buf: &mut &[u8]) -> Result<Self, ProgramError> {
+        Ok(Self::from(Self::fetch_data(buf)?.unwrap_or_default()))
+    }
+
+    // Optimisically write then realloc on failure.
+    // If we serialize and calculate len before realloc we will waste heap space as there is no
+    // free(). Tradeoff between heap usage and compute budget.
+    pub fn store<'a>(
+        &self,
+        account: &AccountInfo<'a>,
+        allow_realloc: bool,
+    ) -> Result<(), ProgramError> {
+        if !account.is_writable || account.executable {
+            return Err(ProgramError::InvalidAccountData);
+        }
+        let realloc_increment = 1024;
+        loop {
+            let mut guard = account.try_borrow_mut_data()?;
+            let data = &mut *guard;
+            let data_len = data.len();
+
+            // Create a new slice so that this new slice
+            // is updated to point to the unwritten data during serialization.
+            // Otherwise, if the account `data` is used directly, `data` will be
+            // the account data itself will be updated to point to unwritten data!
+            let mut writable_data: &mut [u8] = &mut data[..];
+
+            match true
+                .serialize(&mut writable_data)
+                .and_then(|_| self.data.serialize(&mut writable_data))
+            {
+                Ok(_) => break,
+                Err(err) => match err.kind() {
+                    std::io::ErrorKind::WriteZero => {
+                        if !allow_realloc {
+                            return Err(ProgramError::BorshIoError(err.to_string()));
+                        }
+                    }
+                    _ => return Err(ProgramError::BorshIoError(err.to_string())),
+                },
+            };
+            drop(guard);
+            if cfg!(target_os = "solana") {
+                account.realloc(data_len + realloc_increment, false)?;
+            } else {
+                panic!("realloc() is only supported on the SVM");
+            }
+        }
+        Ok(())
+    }
+}
 
 /// Creates associated token account using Program Derived Address for the given seeds.
 /// Required to allow PDAs to be created even if they already have a lamport balance.
