+++ rust/sealevel/programs/ism/test-ism/Cargo.toml
@@ -13,7 +13,6 @@ borsh = "0.9"
 solana-program = "=1.14.13"
 serializable-account-meta = { path = "../../../libraries/serializable-account-meta" }
 hyperlane-core = { path = "../../../hyperlane-core" }
-hyperlane-sealevel-mailbox = { path = "../../mailbox", features = ["no-entrypoint"] }
 hyperlane-sealevel-interchain-security-module-interface = { path = "../../../libraries/interchain-security-module-interface" }
 
 solana-program-test = { version = "=1.14.13", optional = true }
