+++ rust/sealevel/programs/mailbox/src/lib.rs
@@ -1,7 +1,7 @@
 //! Hyperlane Mailbox contract for Sealevel-compatible (Solana Virtual Machine) chains.
 
 #![deny(warnings)]
-// #![deny(missing_docs)] // FIXME
+// #![deny(missing_docs)]
 #![deny(unsafe_code)]
 
 pub mod accounts;
