+++ rust/sealevel/programs/validator-announce/Cargo.toml
@@ -12,7 +12,6 @@ borsh = "0.9"
 solana-program = "=1.14.13"
 ecdsa-signature = { path = "../../libraries/ecdsa-signature" }
 hyperlane-core = { path = "../../hyperlane-core" }
-hyperlane-sealevel-mailbox = { path = "../mailbox", features = ["no-entrypoint"] }
 serializable-account-meta = { path = "../../libraries/serializable-account-meta" }
 thiserror = "1.0"
