+++ rust/sealevel/programs/validator-announce/src/accounts.rs
@@ -1,5 +1,5 @@
+use account_utils::{AccountData, SizedData};
 use borsh::{BorshDeserialize, BorshSerialize};
-use hyperlane_sealevel_mailbox::accounts::{AccountData, SizedData};
 use solana_program::{account_info::AccountInfo, program_error::ProgramError, pubkey::Pubkey};
 
 use crate::validator_announce_pda_seeds;
